// src/enum/type-bank-slip.enum.ts
export enum TypeBankSlipEnum {
  PREFEITURAS = 'PREFEITURAS',
  SANEAMENTO = 'SANEAMENTO',
  ENERGIA_ELETRICA_GAS = 'ENERGIA_ELETRICA_GAS',
  TELECOMUNICACOES = 'TELECOMUNICACOES',
  ORGAOS_GOVERNAMENTAIS = 'ORGAOS_GOVERNAMENTAIS',
  OUTROS = 'OUTROS',
  MULTA_TRANSITO = 'MULTA_TRANSITO',
  BANCO = 'BANCO',
}